% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_0_360_to_0_180.R
\name{.convert_0_360_to_0_180}
\alias{.convert_0_360_to_0_180}
\title{(Internal) Convert Raster Longitude from 0-360 to -180/180}
\usage{
.convert_0_360_to_0_180(raster)
}
\arguments{
\item{raster}{A \code{terra::SpatRaster} object with longitude extent within (0, 360).}
}
\value{
A \code{terra::SpatRaster} object with longitude extent within (-180, 180).
}
\description{
Splits a SpatRaster with longitude range (0 to 360) into two parts, shifts the
(180 to 360) portion to (-180 to 0), and merges them back into a single raster
with standard longitude range (-180 to 180).
}
\keyword{internal}
