% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nc_timestamps.R
\name{get_nc_timestamps}
\alias{get_nc_timestamps}
\title{Extract Timestamps from a NetCDF File}
\usage{
get_nc_timestamps(nc_file, type)
}
\arguments{
\item{nc_file}{Character string. The file path to the NetCDF file.}

\item{type}{Character string. The type of dataset to determine how the time
variable is parsed. Supported options are \code{"modis"}, \code{"chirps"}, \code{"fldas_abh"},
or \code{"other"}.}
}
\value{
A vector of timestamps. Depending on the \code{type}, this may be a \code{Date}
object, \code{POSIXct} object, or character string (for \code{fldas_abh}).
}
\description{
Extracts and formats time variables from a NetCDF file based on the specific
dataset type (e.g., MODIS, CHIRPS, FLDAS).
}
\examples{
\dontrun{
dates <- get_nc_timestamps("path/to/modis_data.nc", type = "modis")
}
}
